requirements.txt
# Core dependencies
streamlit==1.32.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector store and embeddings
faiss-cpu==1.8.0
numpy==1.24.3

# OpenAI
openai==1.12.0

# Text processing
tiktoken==0.6.0

# Additional utilities
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.6
